css: give logo a white background
The logo is placed on a color gradient. And the colors in the gradient
are the same as the colors in the logo. This makes it very hard to
recognize the parts of the logo.

Add a white background to the log to make the logo elements visible in
the titlebar.

Also add a bit more space between the logo and the text. Both are white
so extra background color between the white parts looks nicer to my
eyes.

Signed-off-by: Stefan Lengfeld <stefan.lengfeld@inovex.de>