Avoid posthog for local deployment
We don't want to have the local test deployments be tracked in the
posthog analytics. This waters the numbers.

So add some javascript code to disable capturing for local deployments.
It's easier to do it here than to make filters on posthog.

I have taken the code from

    https://posthog.com/tutorials/multiple-environments#opt-out-of-capturing-based-on-url

but also added support for IPv6.